Implemented allocation selection view + Added missing entities

+++ src/main/java/me/gablerlog/webapp/db/entity/Capacity.java
@@ -3,12 +3,12 @@ package me.gablerlog.webapp.db.entity;
 public class Capacity {
 	
 	private double volume;
-	private double weight;
+	private int	   weight;
 	
 	public Capacity() {
 	}
 	
-	public Capacity(double volume, double weight) {
+	public Capacity(double volume, int weight) {
 		this.volume = volume;
 		this.weight = weight;
 	}
@@ -21,11 +21,11 @@ public class Capacity {
 		this.volume = volume;
 	}
 	
-	public double getWeight() {
+	public int getWeight() {
 		return weight;
 	}
 	
-	public void setWeight(double weight) {
+	public void setWeight(int weight) {
 		this.weight = weight;
 	}
 }
